2008-04-15 13:32 UTC+0100 Viktor Szakats (harbour.01 syenar hu)    * source/rtl/tget.prg      + TGet instance variable ordering in class definition        made CA-Cl*pper compatible, so that it's now possible to        access object vars as array elements the same way.        Read-only access is strongly recommended when using this        unofficial access method, and the practice is overall        discouraged.        NOTE: oGet[8] is not supported in Harbour.        NOTE: in oGet[11] (in C52 mode), oGet[17] (in C53 mode)              only the first char is compatible, which is the              type. The rest is 'trash' in CA-Cl*pper.
   * tests/rto_get.prg
     + Extended tests with regards to array access of the 
       TGet object.

   * source/rtl/tbcolumn.prg
     * Formatting.

   * source/common/hbstr.c
   * source/compiler/ppcomp.c
   * source/compiler/hbgenerr.c
   * utils/hbpp/hbpp.c
     ! Fixed some BCC58 warnings.
     ; TOFIX: These remain:
       Warning W8017 C:\devl\BCC58\Include\stdint.h 77: Redefinition of 'INT16_MIN' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 78: Redefinition of 'INT32_MIN' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 79: Redefinition of 'INT64_MIN' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 82: Redefinition of 'INT16_MAX' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 83: Redefinition of 'INT32_MAX' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 84: Redefinition of 'INT64_MAX' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 87: Redefinition of 'UINT16_MAX' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 88: Redefinition of 'UINT32_MAX' is not identical
       Warning W8017 C:\devl\BCC58\Include\stdint.h 89: Redefinition of 'UINT64_MAX' is not identical
       Warning W8084 source\rtl\hbinet.c 507: Suggest parentheses to clarify precedence in function HB_FUN_HB_INETINIT

   * utils/hbtest/rt_main.h
     ! Typo.

   * utils/hbtest/make_c5x.bat
     + Cleanups, fixes, enhancements.
     * Changed invocation for C53. Now "53" (without quotes) 
       should be used as command line parameter.
     + Now automatically selects linker.
     ! Fixed MSC parameters for C53.
     + Copyright added.

+++ harbour/source/common/hbstr.c
@@ -567,7 +567,7 @@ static BOOL hb_str2number( BOOL fPCode, const char* szNum, ULONG ulLen, HB_LONG 
             {
                *dVal = *dVal * 10.0 + ( c - '0' );
             }
-            else if( *lVal < lLimV || ( *lVal <= lLimV && ( c - '0' ) <= iLimC ) )
+            else if( *lVal < lLimV || ( *lVal <= lLimV && ( ( int ) ( c - '0' ) ) <= iLimC ) )
             {
                *lVal = *lVal * 10 + ( c - '0' );
             }
